refactor: UserVerificationService JpaRepository 의존성 역전으로 끊고 UserRepository 의존하도록 수정
