+++ src/main/java/site/travellaboratory/be/user/presentation/controller/UserProfileReaderController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import site.travellaboratory.be.user.application.UserProfileReaderService;
+import site.travellaboratory.be.user.application.service.UserProfileReaderService;
 import site.travellaboratory.be.common.annotation.UserId;
 import site.travellaboratory.be.user.presentation.response.reader.UserProfileResponse;
